// API.ts
export enum UserType {
  user = "user",
  manager = "manager",
  admin = "admin",
}
// ...
export type User = {
  __typename: "User",
  id: string,
  username: string,
  owner?: string | null,
  type?: UserType | null,
  balance?: number | null,
  createdAt: string,
  updatedAt: string,
};
// ...
export type ModelUserConnection = {
  __typename: "ModelUserConnection",
  items:  Array<User | null >,
  nextToken?: string | null,
};
// ...
export type UsersByTypeAndUsernameQuery = {
  usersByTypeAndUsername?:  {
    __typename: "ModelUserConnection",
    items:  Array< {
      __typename: "User",
      id: string,
      username: string,
      owner?: string | null,
      type?: UserType | null,
      balance?: number | null,
      createdAt: string,
      updatedAt: string,
    } | null >,
    nextToken?: string | null,
  } | null,
};